requirements.txt
fastapi
Pillow
numpy
starlette
uvicorn
uvicorn[standard]
fastapi[standard]
